+++ MOP/src/Misc/CustomExtensions.cs
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.If not, see<http://www.gnu.org/licenses/>.
 
+using System.Collections.Generic;
+
 namespace MOP
 {
     static class CustomExtensions
@@ -27,6 +29,27 @@ namespace MOP
         public static bool ContainsAny(this string lookIn, params string[] lookFor)
         {
             for (int i = 0; i < lookFor.Length; i++)
+            {
+                // Value found? Return true.
+                if (lookIn.Contains(lookFor[i]))
+                {
+                    return true;
+                }
+            }
+
+            // Nothing has been found? Return false.
+            return false;
+        }
+
+        /// <summary>
+        /// Checks if string contains any of the values provided in lookFor.
+        /// </summary>
+        /// <param name="lookIn">String in which we should look for</param>
+        /// <param name="lookFor">Array that we want to look for in lookIn</param>
+        /// <returns></returns>
+        public static bool ContainsAny(this string lookIn, List<string> lookFor)
+        {
+            for (int i = 0; i < lookFor.Count; i++)
             {
                 // Value found? Return true.
                 if (lookIn.Contains(lookFor[i]))